requirements.txt
google-adk
google-generativeai
opencv-python
pillow
python-dotenv
pandas
numpy
